requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dateparser==1.2.0
Deprecated==1.2.18
filelock==3.15.4
Flask==3.1.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gunicorn==23.0.0
huggingface-hub==0.23.4
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.5
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
matplotlib==3.8.3
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
propcache==0.3.2
PyAutoGUI==0.9.54
pycryptodome==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyRect==0.2.0
PyScreeze==1.0.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.4
safetensors==0.4.3
six==1.16.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
